Move the Timer class to a separate module
